menumanager/api/serializers.py
from django.contrib.auth.models import User, Group
from .models import DefaultSandwich, Ingredient, Promotion
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = '__all__'


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = '__all__'


class SandwichSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = DefaultSandwich
        fields = '__all__'


class IngredientSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Ingredient
        fields = '__all__'


class PromotionSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Promotion
        fields = '__all__'