menumanager/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
idna==2.8
pytz==2018.9
sqlparse==0.3.0
urllib3==1.24.1
